requirements.txt
asgiref==3.7.2
astroid==2.15.6
autopep8==2.0.4
beautifulsoup4==4.12.2
colorama==0.4.6
dill==0.3.7
Django==4.2.4
django-adminlte3==0.1.6
django-bootstrap-v5==1.0.11
djangorestframework==3.14.0
flake8==6.1.0
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
Markdown==3.4.4
mccabe==0.7.0
mypy==1.5.1
mypy-extensions==1.0.0
packaging==23.1
platformdirs==3.10.0
pluggy==1.3.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
pylint==2.17.5
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pytest==7.4.0
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3
soupsieve==2.4.1
sqlparse==0.4.4
tomlkit==0.12.1
typing_extensions==4.7.1
tzdata==2023.3
wrapt==1.15.0
